+++ src/Models/dataTypings.ts
@@ -1,0 +1,7 @@
+export interface ISquerMeterData {
+	value: number
+}
+
+export interface IDataBase {
+	squareMeter: ISquerMeterData
+}