+++ src/__tests__/index.ts
@@ -1,5 +1,0 @@
-describe('should be tested', () => {
-	it('test', () => {
-		expect(true).toBe(true)
-	})
-})