+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+-   Model, Database and init some tests with jest and supertest
+
 ## [1.1.0] - 2021-05-07
 
 ### Fixed